+++ CHANGES.rst
@@ -82,6 +82,7 @@ ENHANCEMENTS:
 * Use cucumber "gherkin-languages.json" now (simplify: Gherkin v6 aliases, language usage)
 * Support emojis in ``*.feature`` files and steps
 * Select-by-location: Add support for "Scenario container" (Feature, Rule, ScenarioOutline) (related to: #391)
+* issue #1246: ScenarioOutline: Support template placeholders in ExamplesTable.tags (provided-by: jenisys)
 * pull #1185: Read Examples table data from CSV file (provided-by: rajanmanwansh)
 * issue #1183: Support Background Steps with Placeholders for Scenario Outlines (submitted by: hemendra0851)
 * pull #1097: Support And-Step as initial Scenario step if Background Steps exist (provided-by: aneeshdurg)